@@ -1,5 +1,7 @@
 name: flutter_app_cuberto
 description: A new Flutter application. (cuberto)
+publish_to: none
+
 
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
@@ -14,7 +16,7 @@ description: A new Flutter application. (cuberto)
 version: 1.1.0+0
 
 environment:
-  sdk: ">=2.2.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -22,7 +24,7 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
+  cupertino_icons: ^1.0.4
   cuberto_bottom_bar:
     path: ../
 
