@@ -1,10 +1,11 @@
 name: cuberto_bottom_bar
 description: An animated bottom navigation bar supporting drawer icon at start or end or can be used with the drawer icon and is provided with 2 different styles.
-version: 1.2.0
+version: 2.0.0
 homepage: https://github.com/kushalmahapatro/cuberto_bottom_bar
 
 environment:
-  sdk: ">=2.5.0 <3.0.0"
+    sdk: ">=2.12.0 <3.0.0"
+
 
 dependencies:
   flutter:
