@@ -8,48 +8,48 @@ const int kAnimationDuration = 300;
 
 class CubertoBottomBar extends StatefulWidget {
   /// The callback that will be executed each time the tab is changed
-  final Function(int position, String title, Color tabColor)
+  final Function(int position, String title, Color? tabColor)
       onTabChangedListener;
 
   /// This color is used to show that the tab is inactive i.e not selected
-  final Color inactiveIconColor;
+  final Color? inactiveIconColor;
 
   /// This color is used to specify the tab color
-  final Color tabColor;
+  final Color? tabColor;
 
   /// This color is used to specify the color of text in the tab
-  final Color textColor;
+  final Color? textColor;
 
   /// This color is used to set up the background color of the bottom bar
-  final Color barBackgroundColor;
+  final Color? barBackgroundColor;
 
   /// List of [TabData] to set up the bottom nav bar
   final List<TabData> tabs;
 
   /// [BorderRadius] to add border to the bottom bar.
-  final BorderRadius barBorderRadius;
+  final BorderRadius? barBorderRadius;
 
   /// This int value is used to programmatically change the index of the tabs
   final int selectedTab;
 
   /// To add a navgation drawer of [CubertoDrawer] type
-  final CubertoDrawer drawer;
+  final CubertoDrawer? drawer;
 
   /// [CubertoTabStyle] to be defined as required, by default it will be [CubertoTabStyle.STYLE_NORMAL]
-  final CubertoTabStyle tabStyle;
+  final CubertoTabStyle? tabStyle;
 
   /// The [Key] of the [CubertoBottomBar]
-  final Key key;
+  final Key? key;
 
   /// The inner padding of the [CubertoBottomBar]
-  final EdgeInsets padding;
+  final EdgeInsets? padding;
 
   /// The list of shadows of the [CubertoBottomBar]
-  final List<BoxShadow> barShadow;
+  final List<BoxShadow>? barShadow;
 
   CubertoBottomBar({
-    @required this.tabs,
-    @required this.onTabChangedListener,
+    required this.tabs,
+    required this.onTabChangedListener,
     this.key,
     this.selectedTab = 0,
     this.inactiveIconColor,
@@ -61,9 +61,7 @@ class CubertoBottomBar extends StatefulWidget {
     this.tabStyle,
     this.padding,
     this.barShadow,
-  })  : assert(onTabChangedListener != null),
-        assert(tabs != null),
-        assert(tabs.length > 1 && tabs.length < 5);
+  }) : assert(tabs.length > 1 && tabs.length < 5);
 
   @override
   CubertoBottomBarState createState() => CubertoBottomBarState();
@@ -74,15 +72,15 @@ class CubertoBottomBarState extends State<CubertoBottomBar> {
   IconData activeIcon = Icons.search;
   int currentSelected = 0;
   double _circleAlignX = 0;
-  Color circleColor;
-  Color activeIconColor;
-  Color inactiveIconColor;
-  Color barBackgroundColor;
-  Color textColor;
-  Color tabColor;
-  CubertoDrawerStyle drawerStyle;
-  CubertoTabStyle tabStyle;
-  Icon drawerIcon;
+  Color? circleColor;
+  Color? activeIconColor;
+  Color? inactiveIconColor;
+  Color? barBackgroundColor;
+  Color? textColor;
+  Color? tabColor;
+  CubertoDrawerStyle? drawerStyle;
+  CubertoTabStyle? tabStyle;
+  Icon? drawerIcon;
 
   @override
   void didChangeDependencies() {
@@ -114,13 +112,13 @@ class CubertoBottomBarState extends State<CubertoBottomBar> {
         color: inactiveIconColor,
       );
     } else {
-      if (widget.drawer.style != null)
-        drawerStyle = widget.drawer.style;
+      if (widget.drawer?.style != null)
+        drawerStyle = widget.drawer?.style ?? CubertoDrawerStyle.NO_DRAWER;
       else
         drawerStyle = CubertoDrawerStyle.NO_DRAWER;
 
-      if (widget.drawer.icon != null)
-        drawerIcon = widget.drawer.icon;
+      if (widget.drawer?.icon != null)
+        drawerIcon = widget.drawer?.icon ?? const Icon(Icons.menu);
       else
         drawerIcon = Icon(
           Icons.menu,
@@ -132,11 +130,12 @@ class CubertoBottomBarState extends State<CubertoBottomBar> {
     } else {
       tabStyle = widget.tabStyle;
     }
-    _setSelected(widget.tabs[widget.selectedTab ?? 0].key);
+    _setSelected(widget.tabs[widget.selectedTab].key);
   }
 
-  _setSelected(Key key) {
-    int selected = widget.tabs.indexWhere((tabData) => tabData.key.toString() == key.toString());
+  _setSelected(Key? key) {
+    int selected = widget.tabs
+        .indexWhere((tabData) => tabData.key.toString() == key.toString());
     if (mounted) {
       setState(() {
         currentSelected = selected;
@@ -163,8 +162,9 @@ class CubertoBottomBarState extends State<CubertoBottomBar> {
     // set action as drawerStyle
     if (drawerStyle != CubertoDrawerStyle.NO_DRAWER) {
       actions = IconButton(
-        icon: drawerIcon,
-        onPressed: widget.drawer.style == CubertoDrawerStyle.END_DRAWER
+        icon: drawerIcon ?? const Icon(Icons.menu),
+        onPressed: (widget.drawer?.style ?? CubertoDrawerStyle.NO_DRAWER) ==
+                CubertoDrawerStyle.END_DRAWER
             ? _handleDrawerButtonEnd
             : _handleDrawerButton,
         tooltip: MaterialLocalizations.of(context).openAppDrawerTooltip,
@@ -176,7 +176,7 @@ class CubertoBottomBarState extends State<CubertoBottomBar> {
     }
     return Container(
       padding: widget.padding ??
-          EdgeInsets.symmetric(
+          const EdgeInsets.symmetric(
             horizontal: 12.0,
             vertical: 8.0,
           ),
@@ -193,7 +193,7 @@ class CubertoBottomBarState extends State<CubertoBottomBar> {
             ],
       ),
       child: setUpTabs(
-        drawerStyle,
+        drawerStyle ?? CubertoDrawerStyle.NO_DRAWER,
         widget.tabs,
         widget.onTabChangedListener,
         actions,
@@ -203,7 +203,7 @@ class CubertoBottomBarState extends State<CubertoBottomBar> {
 
   rowTabs(
     List<TabData> tabs,
-    Function(int position, String title, Color tabColor) onTabChangedListener,
+    Function(int position, String title, Color? tabColor) onTabChangedListener,
   ) {
     return SafeArea(
       child: Row(
@@ -211,24 +211,24 @@ class CubertoBottomBarState extends State<CubertoBottomBar> {
         mainAxisAlignment: MainAxisAlignment.spaceEvenly,
         crossAxisAlignment: CrossAxisAlignment.center,
         children: tabs
-            .map((t) =>
-                TabItem(
+            .map((t) => TabItem(
                   key: t.key,
                   tabStyle: tabStyle,
                   selected: t.key == tabs[currentSelected].key,
                   iconData: t.iconData,
                   title: t.title,
-                  iconColor: inactiveIconColor,
-                  textColor: textColor,
+                  iconColor: inactiveIconColor ?? Colors.black,
+                  textColor: textColor ?? Colors.black,
                   backGroundGradientColor: t.tabGradient,
-                  tabColor: t.tabColor == null ? inactiveIconColor : t.tabColor,
+                  tabColor: t.tabColor ?? inactiveIconColor ?? Colors.black,
                   borderRadius: t.borderRadius,
                   callbackFunction: (uniqueKey) {
-                    int selected =
-                        tabs.indexWhere((tabData) => tabData.key.toString() == uniqueKey.toString());
+                    int selected = tabs.indexWhere((tabData) =>
+                        tabData.key.toString() == uniqueKey.toString());
                     _setSelected(uniqueKey);
                     _initAnimationAndStart(_circleAlignX, 1);
-                    onTabChangedListener(selected, t.title, inactiveIconColor);
+                    onTabChangedListener(
+                        selected, t.title, inactiveIconColor ?? Colors.black);
                   },
                 ))
             .toList(),
@@ -239,7 +239,7 @@ class CubertoBottomBarState extends State<CubertoBottomBar> {
   setUpTabs(
     CubertoDrawerStyle drawerStyle,
     List<TabData> tabs,
-    Function(int position, String title, Color tabColor) onTabChangedListener,
+    Function(int position, String title, Color? tabColor) onTabChangedListener,
     Widget actions,
   ) {
     Widget widget;
@@ -288,42 +288,40 @@ class TabData {
   String title;
 
   /// This onclick function to be called when clicked on Tab
-  Function onclick;
+  Function? onclick;
 
   /// This color is set as the background color of the Tab
-  Color tabColor;
+  Color? tabColor;
 
   /// This gradient color is used a the background gradient of the Tab
-  Gradient tabGradient;
+  Gradient? tabGradient;
 
   /// This borderRadius is used as the the borderRadius of the Tab
-  BorderRadius borderRadius;
+  BorderRadius? borderRadius;
 
   /// This key is used as unique value to the Tab (if not set default UniqueKey will be added to the Tab)
-  Key key;
-
+  Key? key;
 
   TabData(
-      {@required this.iconData,
-      @required this.title,
+      {required this.iconData,
+      required this.title,
       this.onclick,
       this.tabColor,
       this.borderRadius,
       this.tabGradient,
-      this.key }) { key = key == null ?  UniqueKey() : key;}
-
-
+      this.key}) {
+    key = key == null ? UniqueKey() : key;
+  }
 }
 
 class CubertoDrawer {
-
   /// This icon will be used as the drawer icon
-  final Icon icon;
+  final Icon? icon;
 
   /// This style will be used to decide where the drawer icon will be placed
   final CubertoDrawerStyle style;
 
-  const CubertoDrawer({this.icon, this.style});
+  const CubertoDrawer({this.icon, this.style: CubertoDrawerStyle.NO_DRAWER});
 }
 
 /// enum to set teh drawer position
