@@ -19,7 +19,7 @@ class TabItem extends StatefulWidget {
   final bool selected;
 
   /// This callbackFunction is called when the Tab is clicked
-  final Function(Key uniqueKey) callbackFunction;
+  final Function(Key? uniqueKey) callbackFunction;
 
   /// This color is used fot the text inside the Tab
   final Color textColor;
@@ -31,10 +31,10 @@ class TabItem extends StatefulWidget {
   final Color tabColor;
 
   /// This gradient is used a the background gradient color of the Tab
-  final Gradient backGroundGradientColor;
+  final Gradient? backGroundGradientColor;
 
   /// This radius is used the borderRadius of the Tab
-  final BorderRadius borderRadius;
+  final BorderRadius? borderRadius;
 
   /// This value is used to YAlignment of the icon of the Tab
   final double iconYAlign = ICON_ON;
@@ -49,22 +49,21 @@ class TabItem extends StatefulWidget {
   final GlobalKey stickyKey = GlobalKey();
 
   /// This style is used to set the tabStyle
-  final CubertoTabStyle tabStyle;
+  final CubertoTabStyle? tabStyle;
 
   TabItem(
-      {@required this.selected,
-      @required this.iconData,
-      @required this.title,
-      @required this.callbackFunction,
-      @required this.textColor,
-      @required this.iconColor,
-      @required this.tabColor,
+      {required this.selected,
+      required this.iconData,
+      required this.title,
+      required this.callbackFunction,
+      required this.textColor,
+      required this.iconColor,
+      required this.tabColor,
       this.borderRadius,
       this.backGroundGradientColor,
-      this.tabStyle,
-      Key key})
+      this.tabStyle: CubertoTabStyle.STYLE_NORMAL,
+      Key? key})
       : super(key: key);
-
 
   @override
   _TabItemState createState() => _TabItemState();
@@ -87,7 +86,10 @@ class _TabItemState extends State<TabItem> {
     } else {
       backGradient = widget.selected
           ? widget.backGroundGradientColor != null
-              ? widget.backGroundGradientColor
+              ? widget.backGroundGradientColor ??
+                  LinearGradient(
+                      colors: [Colors.transparent, Colors.transparent],
+                      stops: [0.0, 0.7])
               : LinearGradient(colors: [
                   widget.tabColor.withOpacity(0.1),
                   widget.tabColor.withOpacity(0.1)
